% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myopia.R
\docType{data}
\name{myopia}
\alias{myopia}
\title{Myopia data}
\format{
\subsection{\code{myopia}}{

A data frame with 479 rows and 2
\describe{
\item{Light}{Level of night light}
\item{Sight}{Level of myopia}
}
}
}
\source{
https://math.montana.edu/courses/s216/#data
}
\usage{
myopia
}
\description{
Data on myopia from Quinn et al
}
\details{
Module 8 Activity 17:  Nearsighted children study (ChildrenLightSight.csv)>
\url{https://math.montana.edu/courses/s216/data/ChildrenLightSight.csv}
}
\keyword{datasets}
